feat: add initial implementation
This commit is the initial commit for the project. At this point, the loader has been proven to work
in one use case outside of this repository. Various aspects of the tooling and packaging still need
work at this point.

BREAKING CHANGE: Before this commit, the project did not exist.