fix(index.ts): prevent stripping of undefined from serialized DocumentNode
JSON.stringify is used to produce a string of the AST. Unfortunately, that meant that descriptions,
which are sometimes undefined, were getting stripped from the AST objects. That unlikely would have
any impact on any users of the package, but out of principle we shouldn't be removing altering the
serialized AST in any way by default. This commit also improves the testing tooling.

BREAKING CHANGE: Previously, the definition property on ASTNodes would be removed if it had a value
of undefined. This change makes it such that the property will exist with a value of undefined. Any
code relying on the number of properties or the existence of the definition key to exist could begin
breaking as a result of this change.