fix: fix lib files
Whoops! The lib files got out of date! This updates those files and adds a CI check to ensure this
doesn't happen again. Unfortunately, this seems breaking, so I'm bumping the major version.

BREAKING CHANGE: The lib is a major version behind. Therefore, it seems like bumping the major
version again is the right thing to do.